+++ tests/test_preprocessing.py
@@ -20,6 +20,7 @@ import neural_network_model  # noqa: E402
 from neural_network_model.model import SETTING  # noqa: E402
 from neural_network_model.process_data import Preprocessing  # noqa: E402
 from neural_network_model.s3 import MyS3  # noqa: E402
+from tests.model import ImageObject, XObjClass, ImageAddressObject  # noqa: E402
 
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
@@ -188,26 +189,6 @@ def test_integrated(_object):
     # _object.train_test_split()
 
 
-class ImageObject:
-    def ImageDataGenerator(self, *args, **kwargs):
-        # print(args, kwargs)
-        return self
-
-    def flow(self, *args, **kwargs):
-        # print(args, kwargs)
-        return []
-
-
-class XObjClass:
-    @staticmethod
-    def shape():
-        return (1, 2, 3)
-
-    def reshape(self, *args, **kwargs):
-        # print(args, kwargs)
-        return self
-
-
 def img_to_array_func(*args, **kwargs):
     # print(args, kwargs)
     return XObjClass
@@ -216,12 +197,6 @@ def img_to_array_func(*args, **kwargs):
 def load_image_func(*args, **kwargs):
     # print(args, kwargs)
     return None
-
-
-class ImageAddressObject:
-    @property
-    def name(self):
-        return "test_image.jpg"
 
 
 def image_dict_object(*args, **kwargs):
@@ -266,3 +241,7 @@ def test_augment_data(mocker, _object):
     )
 
     _object.augment_data(number_of_images_tobe_gen=5)
+
+
+def test_augment_data_2(mocker, _object):
+    ...
